// src/app/dashboard/interfaces/recipes.interface.ts
// Generated by https://quicktype.io

import { Ingredientes } from "./ingredients.interface";

export interface Recipe {
  tipo:              string;
  title:             string;
  preparacion:       string;
  tiempoCoccion:     string;
  grados:            string;
  tiempoElaboracion: string;
  tiempoTotal:       string;
  alt_image?:        string;
}